refactor(notifications): separate ask schema contract
